test : 리팩토링으로 인한 테스트 수정
- 기존 테스트 수정
- Member 도메인으로 테스트 이동
